refactor(frontend): clean up data upload
- Clean up `src/features/data` by organising components into directories.
- Break up `MapObservations` into smaller files.
