+++ frontend-v2/src/features/data/LoadDataStepper.tsx
@@ -9,8 +9,8 @@ import Button from "@mui/material/Button";
 import Typography from "@mui/material/Typography";
 import LoadData from "./LoadData";
 import { useState } from "react";
-import MapObservations from "./MapObservations";
-import MapDosing from "./MapDosing";
+import MapObservations from "./observations/MapObservations";
+import MapDosing from "./dosing/MapDosing";
 import PreviewData from "./PreviewData";
 import { RootState } from "../../app/store";
 import { DatasetRead, useDatasetCsvUpdateMutation } from "../../app/backendApi";
